requirements.txt
# Editar requirements.txt para versão específica
tweepy==4.14.0
requests>=2.31.0
schedule>=1.2.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0